-태창금속 자재관리

+++ src/main/java/com/mes/mesScm/InOut/DTO/SCM_IO.java
@@ -11,7 +11,7 @@ public class SCM_IO {
     private String part_code;
     private String part_name;
     private String work_date;
-    private double qty;
+    private int qty;
     private String spec;
     private String remark;
     private int rownum;
