+++ src/main/java/com/mes/mesScm/Stock/DTO/SCM_STOCK_SUM_DAY.java
@@ -12,10 +12,10 @@ public class SCM_STOCK_SUM_DAY {
     private String part_grp_code2;
     private String part_code;
     private String part_name;
-    private double prev_qty;
-    private double in_qty;
-    private double out_qty;
-    private double qty;
+    private int prev_qty;
+    private int in_qty;
+    private int out_qty;
+    private int qty;
     private String part_grp_name;
     private String part_grp_name2;
     private String unit_name;
